use grpc APIs to interact with blockchain
