+++ relay/stratoschain/msg_builder.go
@@ -4,14 +4,15 @@ import (
 	"math/big"
 
 	sdktypes "github.com/cosmos/cosmos-sdk/types"
-	"github.com/stratosnet/sds/utils/crypto/ed25519"
-
 	"github.com/cosmos/cosmos-sdk/types/bech32"
-	utiltypes "github.com/stratosnet/sds/utils/types"
+
 	"github.com/stratosnet/stratos-chain/types"
 	pottypes "github.com/stratosnet/stratos-chain/x/pot/types"
 	registertypes "github.com/stratosnet/stratos-chain/x/register/types"
 	sdstypes "github.com/stratosnet/stratos-chain/x/sds/types"
+
+	"github.com/stratosnet/sds/utils/crypto/ed25519"
+	utiltypes "github.com/stratosnet/sds/utils/types"
 )
 
 type Traffic struct {
@@ -115,7 +116,7 @@ func BuildUpdateResourceNodeStakeMsg(networkAddr, ownerAddr utiltypes.Address, s
 	return registertypes.NewMsgUpdateResourceNodeStake(
 		networkAddr[:],
 		ownerAddr[:],
-		&coin,
+		coin,
 		incrStake,
 	)
 }
@@ -128,7 +129,7 @@ func BuildUpdateMetaNodeStakeMsg(networkAddr, ownerAddr utiltypes.Address, stake
 	return registertypes.NewMsgUpdateMetaNodeStake(
 		networkAddr[:],
 		ownerAddr[:],
-		&coin,
+		coin,
 		incrStake,
 	)
 }
