+++ relay/stratoschain/types/types.go
@@ -1,0 +1,11 @@
+package types
+
+import (
+	"math/big"
+)
+
+type ResourceNodeState struct {
+	IsActive  uint32
+	Suspended bool
+	Tokens    *big.Int
+}
